[application] Allow creation of storage entries with empty keys.
This effectively reverts https://crrev.com/c/2190480, which was landed
as an attempt to fix http://crbug.com/40688849, but that fix is actually
incorrect, since storage entries with empty keys are valid, despite
being odd.

If the developer intends to cancel the creation of the new entry, they
can just press Esc instead.

This fixes the issue that creating new entries with non-empty keys would
silently delete the entry with the empty key (if it exists).

Fixed: 373703285
Bug: 40688849
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5999196
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>

+++ front_end/panels/application/SharedStorageItemsView.ts
@@ -87,7 +87,7 @@ export namespace SharedStorageItemsDispatcher {
 
   export interface ItemEditedEvent {
     columnIdentifier: string;
-    oldText: string;
+    oldText: string|null;
     newText: string;
   }
 
@@ -241,7 +241,7 @@ export class SharedStorageItemsView extends StorageItemsView {
 
   async #editingCallback(
       editingNode: DataGrid.DataGrid.DataGridNode<Protocol.Storage.SharedStorageEntry>, columnIdentifier: string,
-      oldText: string, newText: string): Promise<void> {
+      oldText: string|null, newText: string): Promise<void> {
     if (columnIdentifier === 'key' && newText === '') {
       // The Shared Storage backend does not currently allow '' as a key, so we only set a new entry with a new key if its new key is nonempty.
       await this.refreshItems();
@@ -249,7 +249,9 @@ export class SharedStorageItemsView extends StorageItemsView {
       return;
     }
     if (columnIdentifier === 'key') {
-      await this.#sharedStorage.deleteEntry(oldText);
+      if (oldText !== null) {
+        await this.#sharedStorage.deleteEntry(oldText);
+      }
       await this.#sharedStorage.setEntry(newText, editingNode.data.value || '', false);
     } else {
       // The Shared Storage backend does not currently allow '' as a key, so we use ' ' as the default key instead.
